ci: add workflow to compare branches/refs (#522)
Current workflows focus on comparing different openvm refs.
This workflow fixes the openvm ref and compares differences in this repo
itself (guest and host changes).

We could combine the two types of comparisons into one workflow with
more options, but I felt it was simpler to separate the two types of
comparisons for now.